feat: separate user and profile routes
Fixes #1272
