+++ app/frontend/src/AppRoutes.tsx
@@ -2,6 +2,7 @@ import PageTitle from "components/PageTitle";
 import TextBody from "components/TextBody";
 import Contribute from "features/Contribute";
 import EditProfilePage from "features/profile/edit/EditProfilePage";
+import UserPage from "features/profile/view/UserPage";
 import React from "react";
 import { Switch } from "react-router-dom";
 
@@ -36,7 +37,7 @@ import {
   connectionsRoute,
   contributeRoute,
   discussionRoute,
-  editUserRoute,
+  editProfileRoute,
   eventsRoute,
   groupRoute,
   guideRoute,
@@ -47,6 +48,7 @@ import {
   newGuideRoute,
   newPlaceRoute,
   placeRoute,
+  profileRoute,
   resetPasswordRoute,
   searchRoute,
   settingsRoute,
@@ -136,11 +138,14 @@ export default function AppRoutes() {
       {
         // PROFILE
       }
-      <AppRoute isPrivate path={editUserRoute}>
+      <AppRoute isPrivate path={editProfileRoute}>
         <EditProfilePage />
       </AppRoute>
-      <AppRoute variant="full-width" isPrivate path={userRoute}>
+      <AppRoute isPrivate path={profileRoute}>
         <ProfilePage />
+      </AppRoute>
+      <AppRoute variant="full-width" isPrivate path={userRoute}>
+        <UserPage />
       </AppRoute>
       <AppRoute isPrivate path={`${connectionsRoute}/:type?`}>
         <ConnectionsPage />
