+++ app/frontend/src/features/dashboard/DashboardBanners.tsx
@@ -7,7 +7,7 @@ import { accountInfoQueryKey } from "queryKeys";
 import React from "react";
 import { useQuery } from "react-query";
 import { Link } from "react-router-dom";
-import { routeToEditUser, settingsRoute } from "routes";
+import { routeToEditProfile, settingsRoute } from "routes";
 import { service } from "service";
 import makeStyles from "utils/makeStyles";
 
@@ -51,7 +51,7 @@ export default function DashboardBanners() {
                 {UPLOAD_PHOTO}
               </Typography>
               <Typography variant="inherit" paragraph>
-                <MuiLink component={Link} to={routeToEditUser()}>
+                <MuiLink component={Link} to={routeToEditProfile()}>
                   {CLICK_HERE_TO_EDIT}
                 </MuiLink>
               </Typography>
