+++ app/frontend/src/features/profile/hooks/useUpdateUserProfile.ts
@@ -1,10 +1,9 @@
 import { useAuthContext } from "features/auth/AuthProvider";
-import useCurrentUser from "features/userQueries/useCurrentUser";
 import { Empty } from "google-protobuf/google/protobuf/empty_pb";
 import { accountInfoQueryKey } from "queryKeys";
 import { useMutation, useQueryClient } from "react-query";
 import { useHistory } from "react-router-dom";
-import { routeToUser } from "routes";
+import { routeToProfile } from "routes";
 import { service, UpdateUserProfileData } from "service/index";
 import { SetMutationError } from "utils/types";
 
@@ -17,7 +16,6 @@ export default function useUpdateUserProfile() {
   const queryClient = useQueryClient();
   const history = useHistory();
   const userId = useAuthContext().authState.userId;
-  const { data: user } = useCurrentUser();
   const {
     mutate: updateUserProfile,
     reset,
@@ -36,11 +34,7 @@ export default function useUpdateUserProfile() {
       onSuccess: () => {
         queryClient.invalidateQueries(["user", userId]);
         queryClient.invalidateQueries(accountInfoQueryKey);
-        if (user) {
-          history.push(routeToUser(user.username, "about"));
-        } else {
-          throw new Error("User is undefined after saving user profile.");
-        }
+        history.push(routeToProfile("about"));
       },
     }
   );
